Exchange/Transfer room done

+++ application/src/main/java/net/furizon/backend/feature/pretix/objects/order/dto/UpdatePretixPositionRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class UpdatePretixPositionRequest {
     private final String order;
 
-    private final int item;
+    private final long item;
 }
